+++ R/conv_funcs.R
@@ -114,7 +114,11 @@ sep_bdr_nl <- function(x){
 }
 
 progress_bar <- function(prop_complete, max_width, msg){
-  pct_l <- paste0(msg,"; ", fmt(prop_complete*100),"% complete.")
+  pct_l <- formatC(prop_complete*100, digits = 5, format = "fg")
+  a <- gsub("\\s", "", pct_l)
+  b <- gsub("\\S", "", pct_l)
+  b <- ifelse(b==""," ", b)
+  pct_l <- paste0(msg,"; ", ifelse(prop_complete == 1, "100%", a),"%", b, "complete.")
   status_width <- max_width - nchar(pct_l)
   bar_width <- as.integer(prop_complete*status_width)
   space_width <- status_width - bar_width